+++ Modules/Office365DSC/DSCResources/MSFT_EXOAcceptedDomain/MSFT_EXOAcceptedDomain.psm1
@@ -10,7 +10,7 @@ function Get-TargetResource
         $Identity,
 
         [Parameter()]
-        [ValidateSet('Authoritative')]
+        [ValidateSet('Authoritative','InternalRelay')]
         [System.String]
         $DomainType = 'Authoritative',
 
@@ -120,7 +120,7 @@ function Set-TargetResource
         $Identity,
 
         [Parameter()]
-        [ValidateSet('Authoritative')]
+        [ValidateSet('Authoritative','InternalRelay')]
         [System.String]
         $DomainType = 'Authoritative',
 
@@ -179,7 +179,7 @@ function Test-TargetResource
         $Identity,
 
         [Parameter()]
-        [ValidateSet('Authoritative')]
+        [ValidateSet('Authoritative','InternalRelay')]
         [System.String]
         $DomainType = 'Authoritative',