+++ Modules/Office365DSC/Modules/O365DSCReverse.psm1
@@ -122,7 +122,7 @@ function Start-O365ConfigurationExtract
         try
         {
             $resourceName = $ResourceModule.Name.Split('.')[0].Replace('MSFT_', '')
-            $currentWorkload = $ResourceName.Substring(0, 2)
+            [array]$currentWorkload = $ResourceName.Substring(0, 2)
             switch ($currentWorkload.ToUpper())
             {
                 'EX'
@@ -167,8 +167,8 @@ function Start-O365ConfigurationExtract
                 }
             }
             if (($null -ne $ComponentsToExtract -and
-                    $ComponentsToExtract.Contains("chck" + $resourceName)) -or
-                $AllComponents -or ($null -ne $Workloads -and $Workloads.Contains($currentWorkload)))
+                    $ComponentsToExtract -contains "chck" + $resourceName) -or
+                $AllComponents -or ($null -ne $Workloads -and $Workloads -contains $currentWorkload))
             {
                 $shouldSkip = $false
                 $usedPlatforms = Get-ResourcePlatformUsage -Resource $resourceName -ResourceModuleFilePath $ResourceModule.FullName